Refactored BragaCryptoTest
